requirements.txt
pjsua2
websocket-client
python-dotenv
openai-agents
deepgram-sdk>=0.3.0
websockets
pyaudio